ReaderAPI - Implemented service to read all files in a certain folder and parse them on ParserAPI

+++ ReaderAPI/ReaderAPI.Integration/Connected Services/ParserAPI/ParserAPI.cs
@@ -268,7 +268,7 @@ namespace ReaderAPI.Integration.ParserAPI
         public string To { get; set; }
 
         [Newtonsoft.Json.JsonProperty("amount", Required = Newtonsoft.Json.Required.DisallowNull, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int Amount { get; set; }
+        public double Amount { get; set; }
 
         [Newtonsoft.Json.JsonProperty("currency", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
         public string Currency { get; set; }
